Move cmd package into internal
